Change the review score type to ushort

+++ database/migrations/2023_12_28_103151_create_reviews_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('building_id')->constrained();
             $table->foreignId('user_id')->constrained();
-            $table->integer('score');
+            $table->unsignedSmallInteger('score');
             $table->text('body');
             $table->timestamps();
         });
